enhance: add 2 more tools. readability is optional standalone. also google longtail search words
