--- wordpress-seo/helper.py
@@ -1,53 +1,0 @@
-
-import os
-import sys
-from datetime import datetime
-import asyncio
-import logging
-import openai
-
-API_KEY = os.getenv("OPENAI_API_KEY")
-client = openai.OpenAI(api_key=API_KEY)
-MODEL = os.getenv("OBOT_DEFAULT_LLM_MODEL", "gpt-4o")
-
-
-def llm_chat_completion(messages, model=MODEL):
-    return client.chat.completions.create(
-        model=model,
-        messages=messages,
-        temperature=0.1,
-        max_tokens=3000
-    )
-
-
-def setup_logger(name):
-    """Setup a logger that writes to sys.stderr. This will eventually show up in GPTScript's debugging logs.
-
-    Args:
-        name (str): The name of the logger.
-
-    Returns:
-        logging.Logger: The logger.
-    """
-    # Create a logger
-    logger = logging.getLogger(name)
-    logger.setLevel(logging.DEBUG)  # Set the logging level
-
-    # Create a stream handler that writes to sys.stderr
-    stderr_handler = logging.StreamHandler(sys.stderr)
-
-    # Create a log formatter
-    formatter = logging.Formatter(
-        "[WordPress Tool Debugging Log]: %(asctime)s - %(name)s - %(levelname)s - %(message)s"
-    )
-    stderr_handler.setFormatter(formatter)
-
-    # Add the handler to the logger
-    logger.addHandler(stderr_handler)
-
-    return logger
-
-
-logger = setup_logger(__name__)
-
-
