+++ wordpress-seo/main.py
@@ -1,8 +1,9 @@
-from scipy import optimize
-from helper import setup_logger
+from tools.helper import setup_logger
 import sys
-from keywords_suggestions import keywords_suggestions
-from content_optimize_metrics import content_optimize_metrics
+from tools.keywords_suggestions import keywords_suggestions_tool
+from tools.keyword_density import keyword_density_metrics_tool
+from tools.readability import readability_metrics_tool
+from tools.long_tail_keywords import google_search_suggestions_tool
 import os
 import json
 
@@ -16,7 +17,7 @@ def main():
     command = sys.argv[1]
     try:
         match command:
-            case "ContentOptimizeMetrics":
+            case "KeywordDensityMetrics":
                 content = os.getenv("CONTENT")
                 primary_keyword = os.getenv("PRIMARY_KEYWORD")
                 if not primary_keyword:
@@ -27,13 +28,22 @@ def main():
                         secondary_keywords = json.loads(secondary_keywords)
                     except json.JSONDecodeError:
                         raise ValueError("SECONDARY_KEYWORDS must be a valid JSON array of strings")
-                content_optimize_metrics(content, primary_keyword, secondary_keywords)
+                res = keyword_density_metrics_tool(content, primary_keyword, secondary_keywords)
             case "KeywordsSuggestions":
                 content = os.getenv("CONTENT")
-                keywords_suggestions(content)
+                res = keywords_suggestions_tool(content)
+            case "ReadabilityMetrics":
+                content = os.getenv("CONTENT")
+                res = readability_metrics_tool(content)
+            case "LongTailKeywords":
+                seed_keyword = os.getenv("SEED_KEYWORD")
+                num_suggestions = int(os.getenv("NUM_SUGGESTIONS", 5))
+                res = google_search_suggestions_tool(seed_keyword, num_suggestions)
             case _:
                 print(f"Unknown command: {command}")
                 sys.exit(1)
+        
+        print(json.dumps(res, indent=4))
     except Exception as e:
         print(f"Running command: {' '.join(sys.argv)} failed. Error: {e}")
         sys.exit(1)
