+++ wordpress-seo/tools/helper.py
@@ -1,0 +1,40 @@
+import sys
+import logging
+
+
+# from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS
+# hardcode it here so we can remove sklearn package
+ENGLISH_STOP_WORDS = frozenset({'should', 'almost', 'hereby', 'except', 'name', 'behind', 'off', 'back', 'with', 'therefore', 'must', 'hers', 'detail', 'fill', 'thereafter', 'out', 'all', 'own', 'top', 'during', 'whereby', 'moreover', 'enough', 'move', 'con', 'of', 'rather', 'although', 'its', 'below', 'among', 'him', 'not', 'please', 'further', 'due', 'six', 'namely', 'under', 'indeed', 'eight', 'made', 'forty', 'seeming', 'whereas', 'cry', 'most', 'many', 'my', 'eleven', 'front', 'sometime', 'next', 'some', 'thereby', 'himself', 'whole', 'along', 'itself', 'i', 'these', 'twenty', 'her', 'down', 'hereafter', 'nobody', 'keep', 'whose', 'may', 'we', 'several', 'while', 'he', 'you', 'which', 'somehow', 'to', 'it', 'within', 'thin', 'anything', 'who', 'un', 'beyond', 'around', 'how', 'yet', 'above', 'sixty', 'in', 'thru', 'yours', 'four', 'two', 'beforehand', 'interest', 'been', 'everyone', 'mine', 'sometimes', 'up', 'then', 'neither', 'ltd', 'cant', 'are', 'sincere', 'whom', 'what', 'has', 'for', 'each', 'themselves', 'she', 'anywhere', 'do', 'found', 'done', 'because', 'thence', 'onto', 'get', 'everywhere', 'besides', 'would', 'also', 'towards', 'but', 'wherein', 'over', 'full', 'them', 'nine', 'the', 'both', 'ours', 'becomes', 'myself', 'anyhow', 'eg', 'our', 'another', 'fifty', 'via', 'across', 'between', 'so', 'against', 'every', 'latterly', 'since', 'where', 'see', 'throughout', 'had', 'other', 'hence', 'your', 'alone', 'amoungst', 'empty', 'me', 'much', 'fire', 'none', 'they', 'yourselves', 'becoming', 'might', 'and', 'seemed', 'us', 'yourself', 'go', 'only', 'beside', 'was', 'often', 'well', 'wherever', 'meanwhile', 'together', 'nowhere', 'part', 'amount', 'hundred', 'herself', 'am', 'per', 'whether', 'ourselves', 'everything', 're', 'can', 'before', 'serious', 'there', 'have', 'either', 'no', 'were', 'five', 'whither', 'is', 'bill', 'however', 'here', 'anyway', 'more', 'few', 'whenever', 'twelve', 'cannot', 'somewhere', 'first', 'thick', 'last', 'co', 'call', 'on', 'same', 'now', 'without', 'this', 'de', 'system', 'side', 'again', 'at', 'someone', 'fifteen', 'otherwise', 'perhaps', 'herein', 'anyone', 'elsewhere', 'ie', 'noone', 'formerly', 'become', 'their', 'an', 'never', 'very', 'be', 'hasnt', 'nothing', 'three', 'being', 'into', 'show', 'whereafter', 'find', 'mill', 'something', 'though', 'until', 'about', 'his', 'once', 'that', 'others', 'etc', 'whoever', 'toward', 'thereupon', 'put', 'already', 'became', 'nevertheless', 'therein', 'couldnt', 'mostly', 'else', 'seems', 'afterwards', 'any', 'one', 'seem', 'third', 'thus', 'describe', 'will', 'as', 'those', 'when', 'ten', 'amongst', 'after', 'always', 'ever', 'than', 'inc', 'former', 'whence', 'least', 'bottom', 'hereupon', 'from', 'give', 'if', 'too', 'such', 'through', 'whatever', 'could', 'by', 'less', 'or', 'even', 'whereupon', 'why', 'nor', 'take', 'upon', 'latter', 'a', 'still'})
+
+
+def setup_logger(name):
+    """Setup a logger that writes to sys.stderr. This will eventually show up in GPTScript's debugging logs.
+
+    Args:
+        name (str): The name of the logger.
+
+    Returns:
+        logging.Logger: The logger.
+    """
+    # Create a logger
+    logger = logging.getLogger(name)
+    logger.setLevel(logging.DEBUG)  # Set the logging level
+
+    # Create a stream handler that writes to sys.stderr
+    stderr_handler = logging.StreamHandler(sys.stderr)
+
+    # Create a log formatter
+    formatter = logging.Formatter(
+        "[WordPress Tool Debugging Log]: %(asctime)s - %(name)s - %(levelname)s - %(message)s"
+    )
+    stderr_handler.setFormatter(formatter)
+
+    # Add the handler to the logger
+    logger.addHandler(stderr_handler)
+
+    return logger
+
+
+logger = setup_logger(__name__)
+
+
